% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.R
\name{cor.dist_boxplot}
\alias{cor.dist_boxplot}
\title{Gene Matrix Correlation Distribution Boxplot}
\usage{
cor.dist_boxplot(cor_list, cor_type, file, ext = "pdf")
}
\arguments{
\item{cor_list}{List of well-to-well correaltion distributions.  Recommended to use the output list(s) from the cor.parser function.}

\item{cor_type}{Correlation type.  Choose from 'pearson', 'spearman' or 'kendall'.}

\item{file}{A file destination/name for the plot output.  File format must be '.pdf'.}
}
\description{
This function is used to create a boxplot of the well-to-well correlation across different groups.
}
\examples{

}
\keyword{Boxplot}
\keyword{Correlation}
\keyword{Gene}
\keyword{Matrix}
