% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.R
\name{cor.parser}
\alias{cor.parser}
\title{Correlation Matrix Parser}
\usage{
cor.parser(cm, metadata = NULL, grouping_var = NULL)
}
\arguments{
\item{cm}{Correlation Matrix.}

\item{metadata}{Metadata table with columns of named features and rows of unique barcodes.}

\item{grouping_var}{Grouping variable(s) used to define groups within correlation matrix.  Must be vector where each element matches a column header in the metadata table (ie. c("Type", "Group"))}
}
\description{
Parse a correlation matrix to find inter and intra-group distributions.  Returns a list with named elements intracor_list & intercor_list
}
\examples{
# Parse a correlation matrix for inter and intra-group distributions.

}
\keyword{Correlation}
\keyword{Gene}
\keyword{Matrix}
