% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.R
\name{cor.stat_table}
\alias{cor.stat_table}
\title{Gene Matrix Correlation Distribution Statistics Table}
\usage{
cor.stat_table(cor_list, cor_dist_obj = FALSE, file)
}
\arguments{
\item{cor_list}{List of well-to-well correaltion distributions. Recommended to use the output list(s) from the cm_parser function.}

\item{file}{Location to save table in .csv format (eg. "path/to/intracor_stats.csv")}
}
\description{
This function is used to create a table of summary statistics for the distribution of well-to-well correlations across different groups.
}
\examples{

}
\keyword{Boxplot}
\keyword{Correlation}
\keyword{Gene}
\keyword{Matrix}
