% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.R
\name{gm.cor_analysis}
\alias{gm.cor_analysis}
\title{Gene Matrix Correlation Analysis}
\usage{
gm.cor_analysis(gm, metadata, cor_type, quant_or_abs = "quant", lo = 0,
  hi = 1, grouping_var = NULL, heatmap = FALSE, dist_boxplot = FALSE,
  dist_table = FALSE, heatmap_cluster = FALSE, heatmap_label = FALSE,
  ext = "pdf", output_dir = getwd(), ...)
}
\arguments{
\item{gm}{Gene matrix.}

\item{metadata}{Metadata table with columns of named features and rows of unique barcodes.}

\item{cor_type}{Correlation type. Choose from 'pearson', 'spearman' or 'kendall'. Defaults to 'pearson'.}

\item{quant_or_abs}{Choose 'quant' for quantile or 'abs' for absolute threshold filtering.}

\item{lo}{Low Filter, can be absolute or quantile based on the 'quant_or_abs' parameter.}

\item{hi}{High Filter, can be absolute or quantile based on the 'quant_or_abs' parameter.}

\item{grouping_var}{Grouping variable(s) used to define plot color scheme.  Must be vector where each element matches a column header in the metadata table (ie. c("Type", "Group"))}

\item{heatmap}{TRUE/FALSE Generate heatmap.  Defaults to FALSE.}

\item{dist_boxplot}{TRUE/FALSE Generate correlation distribution boxplot.  Defaults to FALSE.}

\item{dist_table}{TRUE/FALSE Generate correlation distribution summary statistics table.  Defaults to FALSE.}

\item{heatmap_cluster}{TRUE/FALSE Perform clustering on heatmap.  Defaults to FALSE.}

\item{heatmap_label}{TRUE/FALSE Label rows and columns with sample names on heatmap.  Defaults to FALSE.}

\item{...}{Optional named parameters to be sent to the aheatmap function in the NMF package.}

\item{cor_dist}{TRUE/FALSE. Return 1-sided correlation distribution instead of Correlation Matrix.}
}
\description{
Run correlation analysis.
}
\examples{


}
\keyword{Boxplot}
\keyword{Correlation}
\keyword{Gene}
\keyword{Matrix}
