% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.R
\name{gm.cormatrix}
\alias{gm.cormatrix}
\title{Gene Matrix to Correlation Matrix}
\usage{
gm.cormatrix(gm, cor_type, quant_or_abs = "quant", lo = 0, hi = 1,
  cor_dist = FALSE)
}
\arguments{
\item{gm}{Gene Matrix.}

\item{cor_type}{Correlation type. Choose from 'pearson', 'spearman' or 'kendall'. Defaults to 'pearson'.}

\item{quant_or_abs}{Choose 'quant' for quantile or 'abs' for absolute threshold filtering.}

\item{lo}{Low Filter, can be absolute or quantile based on the 'quant_or_abs' parameter.}

\item{hi}{High Filter, can be absolute or quantile based on the 'quant_or_abs' parameter.}

\item{cor_dist}{TRUE/FALSE. Return 1-sided correlation distribution instead of Correlation Matrix.}
}
\description{
Generate a correlation matrix from a gene matrix.  Optional filtering parameters allow user to set low and high thresholds for correlation.  User can also return a 1-sided distribution of correlation values instead of the entire matrix.
}
\examples{

}
\keyword{Correlation}
\keyword{Gene}
\keyword{Matrix}
