% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tkra_fxns.R
\name{gm.grouper}
\alias{gm.grouper}
\title{gm.grouper}
\usage{
gm.grouper(metadata, grouping_var)
}
\arguments{
\item{metadata}{Metadata table with columns of named features and rows of unique barcodes.}

\item{grouping_var}{Grouping variable(s) used to define plot color scheme.  Must be vector where each element matches a column header in the metadata table (ie. c("Type", "Group"))}
}
\description{
Group variables in metadata file
}
