% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{gm.log}
\alias{gm.log}
\title{Gene Matrix Log Transform}
\usage{
gm.log(gm, log_base = "ln")
}
\arguments{
\item{gm}{Gene matrix.}

\item{log_base}{Logarithmic base. Value must be numeric or 'ln' for natural log. Defaults to 'ln' for natural log.}
}
\description{
Log transforms expression values in a gene matrix.  Adds a pseudocount of 1 to each value in the matrix to avoid undefined log(0).
}
\examples{
# Log10 transform a gene matrix
gm_log <- gm.log(raw_data$gm, 10)
}
\keyword{Gene}
\keyword{Log}
\keyword{Matrix}
\keyword{Transform}
