% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{gm.merger}
\alias{gm.merger}
\title{Gene Matrix Merger}
\usage{
gm.merger(gms, metadata, merge_type)
}
\arguments{
\item{gms}{List of gene matrices.}

\item{metadata}{Matched list of metadata files (genematrix_report.csv files) for each gene matrix in 'gms'.}

\item{merge_type}{Method type to merge gene matrices.  Options are to merge by the 'union' or 'intersect' of expressed genes.  For the union method, non-observed genes will be replaced with zero read counts.}
}
\description{
Merge gene matrix & metadata files into large matrices for downstream analysis.  Merge by either the intersection or union of expressed genes.  Returns a list with named elements "gm" and "metadata".
}
\examples{
# Merge a list of gene matrices into one large gene matrix with the union of
# expressed genes across gene matrices.
}
\keyword{Gene}
\keyword{Matrix}
\keyword{Merge}
