% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{gm.norm}
\alias{gm.norm}
\title{QC Normalization}
\usage{
gm.norm(gm, method, scale = 1e+06, genelengths = NULL)
}
\arguments{
\item{gm}{Gene matrix.}

\item{method}{Normalize to 'tps' (Transcript Per Scale) or 'cc_mc' (cell coverage by median coverage). If 'tps' is chosen, the 'scale' paramter must be set. To normalize each cell to Transcripts Per Million choose method = 'tps', scale = 1e6.}

\item{scale}{The scale with which to normalize read counts when using the 'tps' method. Defaults to Transcripts per Million (TPM), scale = 1e6.}
}
\description{
This function is used to normalize gene matrices.
}
\examples{
# Normalize Gene Matrix to Transcripts per Million (TPM)
}
\keyword{Gene}
\keyword{Matrix}
\keyword{Normalization}
