% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{gm.qc}
\alias{gm.qc}
\title{Gene Matrix Quality Control (gmqc)}
\usage{
gm.qc(gm, metadata = NULL, gm_loc = NULL, gm_qcfilt = TRUE,
  gm_norm = FALSE, gm_norm_method = "tps", gm_norm_scale = 1e+06,
  gm_norm_genelengths = NULL, gm_log = FALSE, gm_log_base = "ln",
  mads = 3, cell_abslowcov = 10000, cells_abslowgenecount = 3,
  gene_cellcount = 1, gene_totcov = 5)
}
\arguments{
\item{gm}{Gene matrix. Optionally, include the raw file location in the 'gm_loc' option.}

\item{gm_loc}{Gene matrix file location (eg. /path/to/genematrix.csv).}

\item{gm_qcfilt}{TRUE/FALSE, perform QC filtering.}

\item{gm_norm}{TRUE/FALSE, normalize gene matrix.}

\item{gm_norm_method}{Normalize to 'tps' (Transcript Per Scale) or 'cc_mc' (cell coverage by median coverage).  If 'tps' is chosen, the 'gm_norm_scale' paramter must be set.  To normalize each cell to Transcripts Per Million choose method = 'tps', gm_norm_scale = 1e6.}

\item{gm_norm_scale}{The scale with which to normalize read counts when using the 'tps' method.  Defaults to Transcripts per Million (TPM), gm_norm_scale = 1e6.}

\item{gm_log}{TRUE/FALSE, perform log transform on gene matrix.}

\item{gm_log_base}{Logrithmic base for transformation.  Choose integer or 'ln' for natural logrithm.  Default is gm_log_base = 'ln'.}

\item{mads}{Number of median absolute deviations from median for QC filtering. Default is mads = 3.}

\item{cell_abslowcov}{Absolute Threshold, set minimum read depth per cell. Defaults to 10,000 reads.}

\item{cells_abslowgenecount}{Absolute Threshold, set minimum gene counts per cell. Defaults to 3 genes}

\item{gene_cellcount}{The number of cells a gene must have > 0 expression to be kept in the gene matrix.  Default is gene_cellcount = 1.}

\item{gene_totcov}{The minimum number of reads for a gene across all cells. Default is gene_totcov = 5.}
}
\description{
This function reads in a gene matrix, filters based on quality control criteria and applies optional normalization and log transformation techniques
}
\examples{
# Perform quality control and normalization on a gene matrix file.
}
\keyword{Gene}
\keyword{Matrix}
\keyword{Normalization}
