% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{gm.reader}
\alias{gm.reader}
\title{Gene Matrix Reader}
\usage{
gm.reader(file_loc, type, ICELL8 = TRUE)
}
\arguments{
\item{file_loc}{File location of gene matrix, or metadata file (eg. '/path/to/genematrix.csv' or '/path/to/genematrix_report.csv').}

\item{type}{File type.  Choose from 'gm' (gene matrix), 'metadata' or 'metadata_header'.}
}
\description{
Reads gene matrix or metadata files from the ICELL8 Primary Pipeline into R
}
\keyword{Gene}
\keyword{Matrix}
\keyword{Reader}
