% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{gm.setup}
\alias{gm.setup}
\title{Gene Matrix Setup}
\usage{
gm.setup(gms_loc, metadata_loc, gm_names = NULL, merge_type = "union",
  transpose = F, ICELL8 = T)
}
\arguments{
\item{gms_loc}{List of gene matrices.}

\item{metadata_loc}{Matched list of metadata files (genematrix_report.csv files) for each gene matrix in 'gms'.}

\item{merge_type}{Method type to merge gene matrices.  Options are to merge by the 'union' or 'intersect' of expressed genes.  For the union method, non-observed genes will be replaced with zero read counts.  Defaults to 'union'.}

\item{transpose}{TRUE/FALSE Transpose gene matrix.  Gene matrix gm[i,j] must have 'i' rows of genes and 'j' columns of samples.}
}
\description{
Reads gene matrix or metadata files from the ICELL8 Primary Pipeline into R and merges multiple output files into a singular object.  Returns a list with named elements 'gm' and 'metadata'.
}
\examples{
# Merge a list of gene matrices into one large gene matrix with the union of
}
\keyword{Gene}
\keyword{Matrix}
\keyword{Merge}
