% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsne.R
\name{gm.tsne}
\alias{gm.tsne}
\title{Gene Matrix tSNE Analysis}
\usage{
gm.tsne(gm, metadata, grouping_var = NULL, dims = 2, initial_dims = 50,
  perplexity, theta = 0.25, max_iter = 1500, pca_center = FALSE,
  pca_scale = FALSE, plot = FALSE, file, ext = FALSE)
}
\arguments{
\item{gm}{Gene Matrix}

\item{metadata}{Metadata table with columns of named features and rows of unique barcodes.}

\item{grouping_var}{Grouping variable(s) used to define perplexity options & plot color scheme.  Must be vector where each element matches a column header in the metadata table (ie. c("Type", "Group"))}

\item{dims}{The number of output dimensions. Defaults to 2.}

\item{initial_dims}{The number of dimensions that should be retained in the initial PCA step. Defaults to 50.}

\item{perplexity}{Perplexity parameter. Defaults to the average number of samples per group divided by the total number of groups (group defined by the 'grouping_var' option).}

\item{theta}{Speed/accuracy trade-off (increase for less accuracy), set to 0.0 for exact tSNE. Defaults to 0.25.}

\item{max_iter}{Number of iterations. Defaults to 1500.}

\item{pca_center}{TRUE/FALSE, Should data be centered before PCA is applied? Defaults to F.}

\item{pca_scale}{TRUE/FALSE, Should data be scaled before PCA is applied? Defaults to F}

\item{plot}{A boolean defining whether or not to create a PCA plot. Defaults to FALSE.}

\item{file}{A file destination/name for the plot output (ie. '/path/to/tsne.pdf').}
}
\description{
Perform t-Distributed Stochastic Neighbor Embedding (tSNE) on a gene matrix.
}
\note{
Defaults for the tSNE parameters have been optimized for general use cases of the ICELL8 platform. These defaults are different from the standard defaults in the Rtsne Package, and may need to be re-optimized for unique applications. Please refer to the original Rtsne package for more information on how to optimize parameters or visit https://distill.pub/2016/misread-tsne/ for more information on optimizing the tSNE analysis.
}
\examples{
# Perform tSNE on the 500 most variable genes
}
\keyword{Gene}
\keyword{Matrix}
\keyword{tSNE}
