% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{gm.var_genes}
\alias{gm.var_genes}
\title{Gene Matrix - Find Highly Variable Genes}
\usage{
gm.var_genes(gm, n_bin, transform_type = "none", lo_mean_cut = 0.1,
  hi_mean_cut = Inf, lo_dsprs_scaled_cut = 1, hi_dsprs_scaled_cut = Inf,
  legend = FALSE, file, ext = "pdf")
}
\arguments{
\item{gm}{Gene Matrix.}

\item{n_bin}{Number of bins from which to calculate scaled variance.  Defaults to 20.}

\item{transform_type}{Transformation type from ("none", "ln" or any integer).  Defaults to "none" for no transformation.  Possible options are "ln" for natural log, or an integer indicating the base of the lograthim used in the transformation (ie. For log2 transformed data, enter 2)}

\item{lo_mean_cut}{Only consider genes highly variable with a mean greater than this number.  Defaults to 0.10.}

\item{hi_mean_cut}{Only consider genes highly variable with a mean less than this number.  Defaults to Inf for Infinity.}

\item{lo_dsprs_scaled_cut}{Only consider genes highly variable if their scaled dispersion is greater than this number.  Defaults to 1.}

\item{hi_dsprs_scaled_cut}{Only consider genes highly variable if their scaled dispersion is less than this number.  Defaults to Inf for Infinity.}

\item{legend}{TRUE/FALSE, if generating the MA plot, include a legend?  Defaults to FALSE.}

\item{file}{A file destination/name for the plot output.  File format must be '.pdf'.  If not provided, plot will be rendered in console.}
}
\description{
This function is used to find highly variable genes.  Genes are seperated into a user-defined number of bins based on their mean expression.  Variance is then scaled for each bin, and the genes with the highest scaled variance are returned.
}
\examples{
# Find highly variable genes, that have mean expression greater than 0.25 and
}
\keyword{Filtering}
\keyword{Gene}
\keyword{Highly}
\keyword{Variable}
