% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn.R
\name{gm.venn}
\alias{gm.venn}
\title{Gene Matrix Venn Diagram}
\usage{
gm.venn(gm, metadata, grouping_var, cell_no = 1, cell_thresh = 0,
  raw_or_percent = c("raw", "percent"), file, ext = "pdf", col = NULL)
}
\arguments{
\item{gm}{Gene Matrix.}

\item{metadata}{Metadata table with columns of named features and rows of unique barcodes.}

\item{grouping_var}{Grouping variable(s) used to define groups within correlation matrix.  Must be vector where each element matches a column header in the metadata table (ie. c("Type", "Group"))}

\item{cell_no}{Cell Number.  The minimum number of cells a gene must be expressed in to be counted within the specified group.  Defaults to 1.}

\item{cell_thresh}{Cell threshold.  The minimum absolute expression level required for a gene to be counted within the specified group.  Defaults to 0 (ie. All expressed genes).}

\item{raw_or_percent}{Choose between c("raw"), c("percent"), or c("raw", "percent"), to display the given statistic type on the venn diagram.  Defaults to c("raw", "percent").}

\item{file}{A file destination/name for the plot output.  File format must be '.pdf'.}
}
\description{
Generate a venn diagram of gene counts for groups in a gene matrix.
}
\examples{
# Generate a venn diagram from a gene matrix, by cell type.
gm.venn(gm, metadata = metadata, grouping_var = c("Type"),
       file = 'gm_venn.pdf')

# Generate a venn diagram from a gene matrix by cell type, only count genes
# that are expressed in 3 or more cells per group and have expression levels
# greater than or equal to 1.
gm.venn(gm, metadata = metadata, grouping_var = c("Type"), cell_no = 3,
        cell_thresh = 1, file = 'gm_venn.pdf')

}
\keyword{Diagram}
\keyword{Gene}
\keyword{Matrix}
\keyword{Venn}
