% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icell8.R
\name{ICELL8.analysis}
\alias{ICELL8.analysis}
\title{ICELL8 Analysis}
\usage{
ICELL8.analysis(ICELL8_data_loc = NULL, gm_file_list = NULL,
  gm_report_list = NULL, cell_type = "all", correlation_type = "pearson",
  parallel_cores = 1, names = NULL, cor_low_thresh = 0,
  cor_high_thresh = 1, quant_or_abs = "quant", gm_norm = FALSE,
  gm_norm_method = "tps", gm_norm_scale = 1e+06,
  gm_norm_genelengths = NULL, gm_log = FALSE, gm_log_base = "ln",
  qc_cell_abslowcov = 10000, qc_cells_abslowgenecount = 3,
  qc_gene_cellcount = 1, qc_gene_totcov = 5, grouping_var = NULL,
  pca_filt_method = "top_var", pca_thresh_cut = 0, pca_top_genes = 500,
  merge_method = "union", correlation_analysis = TRUE,
  cluster_analysis = TRUE, report = TRUE, ICELL8_data = TRUE,
  figure_type = "png", transpose = FALSE, verbose = FALSE,
  output_dir = "ICELL8_output", title = "ICELL8 Analysis Report",
  author = "")
}
\arguments{
\item{ICELL8_data_loc}{Full path to ICELL8_data folder, obtained from primary pipeline (ie. /path/to/ICELL8_data).}

\item{gm_file_list}{Plain text file with a list of gene matrices for analysis.  Each line of the file must include the full path to the gene matrix (ie. /path/to/genematrix.csv).}

\item{gm_report_list}{Plain text file with a list of report files for analysis.  Each line of the must include the full path to the report file for each gene matrix in the --gm_file_list option (ie. /path/to/genematrix_report.csv).}

\item{cell_type}{Optional, comma separated list of cell types/groups of interest. (ie. –t “Positive_Control, K562”).  Defaults to all cell types found in the 'Type' Column of the report files under the --gm_report_list option.}

\item{correlation_type}{Correlation method type, pick from pearson, spearman or kendall.  Defaults to ‘pearson’.}

\item{parallel_cores}{The number of cores to run in parallel.  If running large datasets on a server, increasing the number of parallel cores can dramatically increase the speed of the analysis.}

\item{names}{Optional, comma separated list of names for each gene matrix in the file supplied in the --gm_file_list option (ie. –-names “group1, group2, group3”).  If not specified the basename of each gene matrix file will be used as the name.  For example, /analyses/raw_data/human/A375.100K.withUMI.csv would be shortened to A375.100K.withUMI.}

\item{cor_low_thresh}{For correlation analysis, filter genes from each cell with expression less than this value.}

\item{cor_high_thresh}{For correlation analysis, filter genes from each cell with expression higher than this value.}

\item{quant_or_abs}{Choose 'quant' or 'abs' to define whether the values for the --lo_thresh and --hi_thresh options are quantile or absolute cutoffs.}

\item{gm_norm}{TRUE/FALSE, should gene matrix be normalized?  Defaults to FALSE.  If TRUE, defaults to transcripts per million unless the --gm_norm_method option is also specified.}

\item{gm_norm_method}{What normalization method to use? Normalize to 'tps' (Transcript Per Scale), 'cc_mc' (cell coverage by median coverage) or 'rpkm' (reads per kilobase million). If 'tps' is chosen, the --gm_norm_scale option must be set.  To normalize each cell to Transcripts Per Million set --gm_norm_method to 'tps' and --gm_norm_scale to 1e6.}

\item{gm_norm_scale}{What scaling factor should be used for normalization?  Defaults to 1e6 for Transcripts per Million.}

\item{gm_norm_genelengths}{Provide a table of gene lengths used for rpkm normalization.  Defaults to NULL.}

\item{gm_log}{TRUE/FALSE, should gene matrix be log transformed?  If TRUE, defaults to natural log 'ln', unless the -–gm_log_base option is specified.}

\item{gm_log_base}{What log base should be used for log transformation?  For natural log select ‘ln’.  Defaults to 'ln'.}

\item{qc_cell_abslowcov}{Define minimum read depth per cell. Defaults to 10,000 reads.}

\item{qc_cells_abslowgenecount}{Define minimum gene counts per cell. Defaults to 3 genes.}

\item{qc_gene_cellcount}{The number of cells a gene must have > 0 expression to be kept in the gene matrix.  Defaults to 1.}

\item{qc_gene_totcov}{The minimum number of reads for a gene across all cells. Defaults to 5.}

\item{grouping_var}{Comma separated list of factors from report files to create groups for analysis. Defaults to ‘Group’, which classifies groups by each gene matrix that is entered.  For example, if the analysis should observe differences between cell types and groups enter \"Group, Type\".}

\item{pca_filt_method}{Method to choose which genes will be used for clustering analysis.  Choose from 'quant_exp' (Filter genes below a quantile value specified in the --pca_thresh_cut option), 'top_var' (Select the genes with the highest variance- must also set the --pca_top_genes option), or 'top_exp' (Select the highest expressing genes- must also set the --pca_top_genes option).  If the option is not specified, all genes are used in the analysis.}

\item{pca_thresh_cut}{For clustering analysis, filter genes with expression less than this quantile value (values must be between 0 & 1). Defaults to 0.0 (No genes filtered).}

\item{pca_top_genes}{Select the number of genes to be used in the clustering analysis?  Defaults to 500.}

\item{merge_method}{How should multiple gene matrices be merged?  Choose from ‘intersect’ or ‘union’.  Defaults to ‘union’, where missing genes are filled in with zero read counts.}

\item{correlation_analysis}{TRUE/FALSE, perform correlation analysis. Defaults to TRUE.}

\item{cluster_analysis}{TRUE/FALSE, perform clustering analysis (PCA & tSNE). Defaults to TRUE.}

\item{report}{TRUE/FALSE, generate final report.}

\item{ICELL8_data}{TRUE/FALSE, is data derived from ICELL8 primary pipeline?}

\item{figure_type}{Enter 'pdf' or 'png' for figure outputs.  Defaults to 'pdf'}

\item{transpose}{TRUE/FALSE, transpose gene matrix. Should be columns of samples, rows of genes. Defaults to FALSE.}

\item{verbose}{TRUE/FALSE, print status to console?  Defaults to TRUE.}

\item{output_dir}{Name of the output directory.  Defaults to ICELL8_output in current working directory.  The script will not overwrite existing folders, with the exception of the ICELL8_output folder.}

\item{title}{Title of experiment.}

\item{author}{Name of indvidual(s) performing analysis.}
}
\description{
Perform quality control, normalization, transformation & filtering, correlation and clustering analysis on single cell RNAseq data from ICELL8.  Produce a standardized report.
}
\keyword{ICELL8}
\keyword{analysis}
\keyword{cell}
\keyword{clustering}
\keyword{correlation}
\keyword{pca}
\keyword{rna}
\keyword{rna-seq}
\keyword{rnaseq}
\keyword{seq}
\keyword{single}
\keyword{single-cell}
\keyword{takara}
\keyword{tertiary}
\keyword{tsne}
