% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tkra_fxns.R
\name{lib_load}
\alias{lib_load}
\title{Library Load}
\usage{
lib_load(libs)
}
\arguments{
\item{libs}{List of libraries for loading.}
}
\description{
This function suppresses output when loading libraries.
}
\keyword{General}
\keyword{utility}
