% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tkra_fxns.R
\name{na.omit_list}
\alias{na.omit_list}
\title{Remove NA's from Lists}
\usage{
na.omit_list(y)
}
\arguments{
\item{y}{List.}
}
\description{
This function removes NA's from lists
}
\keyword{General}
\keyword{utility}
