% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca.count}
\alias{pca.count}
\title{PCA Count}
\usage{
pca.count(pca_obj, perc_var = 0.95)
}
\arguments{
\item{pca_obj}{A Principal Component Analysis object from the prcomp function in the 'stats' package.}

\item{perc_var}{The number of Principal Components required to explain 'per_var' (\% Variance) of the data.  Value should be between 0.0 and 1.0, defaults to 0.95 for 95\% of variance.}
}
\description{
This function is used to estimate the number of Principal Components that explain a user-defined percentage of the variance in the data.
}
\examples{
# Find the number of principal components that explain 95\\\% of the data
}
\keyword{Count}
\keyword{PCA}
