% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca.gene_filter}
\alias{pca.gene_filter}
\title{PCA Gene Filter}
\usage{
pca.gene_filter(gm, method, thresh_cut = 0, top_genes = ncol(gm),
  transform_type = "none")
}
\arguments{
\item{gm}{Gene Matrix.}

\item{method}{Filtering method. Choose from from 'quant_exp' (Filter genes below quantile value), 'top_var' (Retain genes with highest scaled variance), or 'top_exp (Retain highest expressing genes).}

\item{thresh_cut}{Quantile filtering threshold if using 'quant_exp' method.  Value should be between 0.0 and 1.0.  Defaults to 0 (all genes retained).}

\item{top_genes}{Number of genes to retain if using 'top_var' or 'top_exp' methods.  Defaults to the number of columns in the gene matrix (all genes retained).}
}
\description{
This function is used to filter genes from a gene matrix as input for Principal Component Analysis
}
\examples{

}
\keyword{Filter}
\keyword{PCA}
