% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca.plot}
\alias{pca.plot}
\title{PCA Plot}
\usage{
pca.plot(pca_obj, gm = NULL, metadata, grouping_var = NULL,
  gene_highlight = c(), pcs = 4, file, ext = "pdf", marker_size = 0.4)
}
\arguments{
\item{pca_obj}{A Principal Component Analysis object from the prcomp function in the 'stats' package.}

\item{metadata}{Metadata table with columns of named features and rows of unique barcodes.}

\item{grouping_var}{Grouping variable(s) used to define plot color scheme.  Must be vector where each element matches a column header in the metadata table (ie. c("Type", "Group"))}

\item{pcs}{The number of Principal Components to plot.  Values must be between 2 & 4.  Defaults to 4.}

\item{file}{A file destination/name for the plot output.  File format must be '.pdf'.  If not provided, plot will be rendered in console.}
}
\description{
This function is used to plot the first x Prinicipal Components (max of 4).
}
\examples{
# Create a 4 x 4 plot of the of the first 4 Principal Components
}
\keyword{PCA}
\keyword{Plot}
