% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{qc.flag_cells_lowcov}
\alias{qc.flag_cells_lowcov}
\title{QC Low Cell Coverage Flagger}
\usage{
qc.flag_cells_lowcov(gm, n, abs_thresh = -Inf)
}
\arguments{
\item{gm}{Gene matrix.}

\item{n}{Number of median absolute deviations from median.}

\item{abs_thresh}{Absolute Threshold, set minimum read depth per cell. Defaults to -Inf reads.}
}
\description{
Flag cells to drop with low overall coverage.
}
\examples{
# Flag cells to drop that are lower than 3 Median Absolute Deviations from
}
\keyword{Cell}
\keyword{Control}
\keyword{Coverage}
\keyword{Dropper}
\keyword{Gene}
\keyword{Matrix}
\keyword{Quality}
