% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{qc.flag_cells_lowgenecount}
\alias{qc.flag_cells_lowgenecount}
\title{QC Low Total Gene Count Flagger}
\usage{
qc.flag_cells_lowgenecount(gm, n, abs_thresh = -Inf)
}
\arguments{
\item{gm}{Gene matrix.}

\item{n}{Number of median absolute deviations from median.}

\item{abs_thresh}{Absolute Threshold, set minimum gene counts per cell. Defaults to -Inf genes.}
}
\description{
This function is used to flag cells to drop with too few total genes expressed.
}
\examples{
# Flag cells to drop that have less than 3 Median Absolute Deviations from
# median total gene count
}
\keyword{Control}
\keyword{Count}
\keyword{Gene}
\keyword{Low}
\keyword{Matrix}
\keyword{Quality}
\keyword{Total}
