% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{qc.flag_genes_lowcellcount}
\alias{qc.flag_genes_lowcellcount}
\title{QC Low Total Cell Count Flagger}
\usage{
qc.flag_genes_lowcellcount(gm, cell_count)
}
\arguments{
\item{gm}{Gene matrix}

\item{cell_count}{The required number of cells in which a given gene has > 0 expression.  If a gene is expressed in less than the number of cells defined in cell_count, the gene is flagged to be dropped from the matrix.}
}
\description{
This function is used to flag genes to drop that are found in too few number of cells.
}
\examples{

}
\keyword{Cell}
\keyword{Control}
\keyword{Count}
\keyword{Gene}
\keyword{Low}
\keyword{Matrix}
\keyword{Quality}
\keyword{Total}
