% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{qc.flag_genes_lowcov}
\alias{qc.flag_genes_lowcov}
\title{QC Low Gene Expression Flagger}
\usage{
qc.flag_genes_lowcov(gm, exp_cutoff)
}
\arguments{
\item{gm}{Gene matrix.}

\item{exp_cutoff}{Flag genes to drop that have < than the defined expression cutoff.}
}
\description{
This function is used to flag genes to drop that have low overall expression.
}
\examples{

}
\keyword{Control}
\keyword{Expression}
\keyword{Gene}
\keyword{Genes}
\keyword{Low}
\keyword{Matrix}
\keyword{Quality}
