% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{qc.mad}
\alias{qc.mad}
\title{QC Median Absolute Deviations}
\usage{
qc.mad(v, n)
}
\arguments{
\item{v}{Vector of values (eg. vector of gene coverage per cell)}

\item{n}{Number of median absolute deviations from median.}
}
\description{
Calculates lower and upper bounds of 'n' median absolute deviations from median of vector.  Used to find outlier cells or genes.  Returns named list with elements 'lo' & 'hi'.
}
\examples{
# Calculate the threshold for dropping cells with lower than 3 Median
# Absolute Deviations from median overall coverage.
}
\keyword{Absolute}
\keyword{Control}
\keyword{Deviation}
\keyword{Gene}
\keyword{Matrix}
\keyword{Median}
\keyword{Quality}
