% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsne.R
\name{tsne.plot}
\alias{tsne.plot}
\title{tSNE Plot}
\usage{
tsne.plot(tsne_obj, gm = NULL, metadata, grouping_var = NULL,
  gene_highlight = NULL, file, gene_highlight_col = c("antiquewhite",
  "firebrick3"), numeric_gradient_col = c("blue", "cyan", "yellow", "red"),
  col_override = NULL, ext = "pdf", marker_size = 1)
}
\arguments{
\item{tsne_obj}{A t-Distributed Stochastic Neighbor Embedding (tSNE) object from the Rtsne function in the 'Rtsne' package.}

\item{metadata}{Metadata table with columns of named features and rows of unique barcodes.}

\item{grouping_var}{Grouping variable(s) used to define plot color scheme.  Must be vector where each element matches a column header in the metadata table (ie. c("Type", "Group"))}

\item{file}{A file destination/name for the plot output.  File format must be '.pdf'.  If not provided, plot will be rendered in console.}
}
\description{
Plot the first two dimensions of a tSNE object.
}
\examples{
# Create a tSNE plot
}
\keyword{Plot}
\keyword{tSNE}
