% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tkra_fxns.R
\name{tsne.shiny}
\alias{tsne.shiny}
\title{Run tsne}
\usage{
tsne.shiny(cluster_data = NULL, gm = NULL, gm_top = NULL,
  metadata = NULL, tsne_obj = NULL, app_loc = file.path(.libPaths(),
  "icell8", "shiny_apps", "tsne"), ...)
}
\description{
Run tsne shiny
}
